Automate gpbackup tests with GH Actions (#1)
Automate gpbackup tests with GH Actions (#1)

This patch adds a corresponding job to run gpbackup integration and end-to-end
tests inside Greengage 6 Docker container. We'll add run with Greengage 7 image
to the matrix when such image is provided. Symlinks to the greengage_path file
should be removed further patches with rebranding.

Ticket: ADBDEV-8302