Disable test reruns when PRs are updated (#6)
Currently, tests are rerun every time a PR description is updated, as well as
when the PR is set to ready for review. To avoid this, remove edited and
ready_for_review from the types in the pipeline settings. By default, a
workflow only runs when a pull_request event's activity type is opened,
synchronized, or reopened. Therefore, we can simply remove all current types.

Ticket: ADBDEV-8616