Fix resize subset restore with replicated table (#5)
Subset resize restore of a replicated table was passed to the plugin for all
batches. As a result, for the second batch, the plugin sent the replicated
table to the helper a second time, while the helper had already moved on to
processing the next table. This resulted in the replicated table data being
restored to a different table. Replicated tables need to be filtered and passed
to the plugin only for the first batch. Subset restore support has also been
added to example_plugin.bash.

Ticket: ADBDEV-8604